#47 refactor: 인터페이스 제거
