+++ src/main/java/com/gregori/category/service/CategoryService.java
@@ -2,14 +2,52 @@ package com.gregori.category.service;
 
 import java.util.List;
 
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import com.gregori.category.domain.Category;
+import com.gregori.category.mapper.CategoryMapper;
 import com.gregori.common.exception.NotFoundException;
 
-public interface CategoryService {
+import lombok.RequiredArgsConstructor;
 
-	Long saveCategory(String name);
-	void updateCategoryName(Long categoryId, String name) throws NotFoundException;
-	void deleteCategory(Long categoryId);
-	Category getCategory(Long categoryId) throws NotFoundException;
-	List<Category> getCategories(int page);
+@Service
+@RequiredArgsConstructor
+public class CategoryService {
+
+	private final CategoryMapper categoryMapper;
+
+	public Long saveCategory(String name) {
+
+		Category category = new Category(name);
+		categoryMapper.insert(category);
+
+		return category.getId();
+	}
+
+	public void updateCategoryName(Long categoryId, String name) throws NotFoundException {
+
+		Category category = categoryMapper.findById(categoryId).orElseThrow(NotFoundException::new);
+		category.updateCategoryName(name);
+		categoryMapper.updateName(category);
+	}
+
+	@Transactional
+	public void deleteCategory(Long categoryId) {
+
+		categoryMapper.deleteById(categoryId);
+	}
+
+	public Category getCategory(Long categoryId) throws NotFoundException {
+
+		return categoryMapper.findById(categoryId).orElseThrow(NotFoundException::new);
+	}
+
+	public List<Category> getCategories(int page) {
+
+		int limit = 10;
+		int offset = (page - 1) * limit;
+
+		return categoryMapper.find(limit, offset);
+	}
 }
